requirements.txt
datasketch==1.5.7
Flask==2.0.3
nltk==3.6.7
python-dotenv==0.20.0
waitress==1.4.4
